+++ tests/integration_tests/integration_test_config_example.json
@@ -225,6 +225,7 @@
                 }
             ],
             "features": {
+                "parallelToolCallsSupported": false,
                 "responseFormatJsonSchemaSupported": false
             }
         },
